Fixed error with last round not always having a 3rd throw. Also tested the Scores data with Doubles. Both Decimals and Doubles returned false as a JSON response..

+++ BowlingPoints/BowlingTest.cs
@@ -76,7 +76,7 @@ namespace BowlingPoints
             BowlingScoresData bsd = new BowlingScoresData();
             ScoreCalculator sc = new ScoreCalculator(bpd.points);
             //bsd.scores = sc.scores;
-            bsd.scores = new List<decimal>();
+            bsd.scores = new List<Double>();
             foreach (int i in sc.scores)
             {
                 bsd.scores.Add(i);
@@ -154,7 +154,7 @@ namespace BowlingPoints
             ScoreCalculator sc = new ScoreCalculator(leest); //now the scoreCalculator can work with the points.
             BowlingScoresData bsd = new BowlingScoresData(); //this is used to grab the scores, once they are calculated.
             //bsd.scores = sc.scores; //Here the scores get grapped.
-            bsd.scores = new List<decimal>();
+            bsd.scores = new List<Double>();
             foreach (int i in sc.scores)
             {
                 bsd.scores.Add(i);
@@ -191,13 +191,13 @@ namespace BowlingPoints
     internal class BowlingScoresData //used when sending back the scores matching a token that matches a game, using POST.
     {
         //The 2 data types needed for the POST request
-        public List<decimal> scores { get; set; }
+        public List<Double> scores { get; set; }
         public string token { get; set; }
 
         internal void WriteToConsole() //This method will print the data to the screen
         {
             Console.Write("scores: [");
-            foreach (decimal i in scores)
+            foreach (Double i in scores)
             {
                 Console.Write(i + " ");
             }
